+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 Brush uses [semantic versioning](https://semver.org/).
 
+## 1.4.2 - 2023-05-24
+
+* Fixed `repository` value in package _README.md_ files
+
+## 1.4.1 - 2023-05-24
+
+* Fixed _README.md_
+
+## 1.4.0 - 2023-05-24
+
+* Updated _README.md_
+* Fixed version
+
+## 1.3.0 - 2023-05-18
+
+* Removed package _.npmrc_ files
+* Removed `canvas-library-tools` package
+
 ## 1.2.1. - 2023-05-11
 
 * Fixed ESLint extend config